+++ docs/app/components/content/ComponentsList.vue
@@ -10,6 +10,7 @@ const { data: components } = await useAsyncData(`components-${props.category}`, 
     .where('path', 'LIKE', '/docs/components/%')
     .where('extension', '=', 'md')
     .where('category', '=', props.category)
+    .where('index', 'IS NULL')
     .select('path', 'title', 'description')
     .all()
 })